% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwst.R
\name{jwst}
\alias{jwst}
\title{jwst palette generator}
\usage{
jwst(name, n)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{carina},
\code{smacs},
\code{souther_nebula_1},
\code{souther_nebula_2},
\code{stephans_quintet}}

\item{n}{Number of colors desired. All palettes have 9 colors. Most color
schemes are derived from If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
jwst palette generator
}
\examples{
jwst("carina")
}
\keyword{colors}
