% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwst.R
\docType{data}
\name{jwst_palettes}
\alias{jwst_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
jwst_palettes
}
\description{
Use \code{\link{jwst}} to construct palettes of desired length.
}
\keyword{datasets}
